stuck post ke database
